split useBattle
